Move `SwiftUIRippleView` to its own file. Move `CGPoint+Distance` extension to its own file.
